ERG-#26 - extract mutation types to separated files in each store module
