+++ store/alerts/mutation-types.js
@@ -1,0 +1,4 @@
+export default {
+    APPEND_ALERT: 'APPEND_ALERT',
+    REMOVE_ALERT: 'REMOVE_ALERT',
+};
