+++ store/attribute/mutation-types.js
@@ -1,0 +1,18 @@
+export default {
+    SET_ATTRIBUTE_ID: 'SET_ATTRIBUTE_ID',
+    INITIALIZE_OPTION_KEYS: 'INITIALIZE_OPTION_KEYS',
+    ADD_ATTRIBUTE_OPTION_KEY: 'ADD_ATTRIBUTE_OPTION_KEY',
+    REMOVE_ATTRIBUTE_OPTION_KEY: 'REMOVE_ATTRIBUTE_OPTION_KEY',
+    REMOVE_ATTRIBUTE_OPTION_VALUE: 'REMOVE_ATTRIBUTE_OPTION_VALUE',
+    SET_ATTRIBUTE_OPTION_KEY: 'SET_ATTRIBUTE_OPTION_KEY',
+    SET_MULTILINGUAL_ATTRIBUTE: 'SET_MULTILINGUAL_ATTRIBUTE',
+    INITIALIZE_OPTION_VALUE_FOR_LANGUAGE_CODE: 'INITIALIZE_OPTION_VALUE_FOR_LANGUAGE_CODE',
+    INITIALIZE_OPTION_VALUES: 'INITIALIZE_OPTION_VALUES',
+    SET_OPTION_VALUE_FOR_LANGUAGE_CODE: 'SET_OPTION_VALUE_FOR_LANGUAGE_CODE',
+    SET_OPTION_VALUE: 'SET_OPTION_VALUE',
+    SET_ATTRIBUTE_CODE: 'SET_ATTRIBUTE_CODE',
+    SET_ATTRIBUTE_PARAMETER: 'SET_ATTRIBUTE_PARAMETER',
+    SET_ATTRIBUTE_GROUPS: 'SET_ATTRIBUTE_GROUPS',
+    SET_ATTRIBUTE_TYPE: 'SET_ATTRIBUTE_TYPE',
+    CLEAR_STATE: 'CLEAR_STATE',
+};
