+++ store/attribute/mutations.js
@@ -2,24 +2,7 @@
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
-export const types = {
-    SET_ATTRIBUTE_ID: 'SET_ATTRIBUTE_ID',
-    INITIALIZE_OPTION_KEYS: 'INITIALIZE_OPTION_KEYS',
-    ADD_ATTRIBUTE_OPTION_KEY: 'ADD_ATTRIBUTE_OPTION_KEY',
-    REMOVE_ATTRIBUTE_OPTION_KEY: 'REMOVE_ATTRIBUTE_OPTION_KEY',
-    REMOVE_ATTRIBUTE_OPTION_VALUE: 'REMOVE_ATTRIBUTE_OPTION_VALUE',
-    SET_ATTRIBUTE_OPTION_KEY: 'SET_ATTRIBUTE_OPTION_KEY',
-    SET_MULTILINGUAL_ATTRIBUTE: 'SET_MULTILINGUAL_ATTRIBUTE',
-    INITIALIZE_OPTION_VALUE_FOR_LANGUAGE_CODE: 'INITIALIZE_OPTION_VALUE_FOR_LANGUAGE_CODE',
-    INITIALIZE_OPTION_VALUES: 'INITIALIZE_OPTION_VALUES',
-    SET_OPTION_VALUE_FOR_LANGUAGE_CODE: 'SET_OPTION_VALUE_FOR_LANGUAGE_CODE',
-    SET_OPTION_VALUE: 'SET_OPTION_VALUE',
-    SET_ATTRIBUTE_CODE: 'SET_ATTRIBUTE_CODE',
-    SET_ATTRIBUTE_PARAMETER: 'SET_ATTRIBUTE_PARAMETER',
-    SET_ATTRIBUTE_GROUPS: 'SET_ATTRIBUTE_GROUPS',
-    SET_ATTRIBUTE_TYPE: 'SET_ATTRIBUTE_TYPE',
-    CLEAR_STATE: 'CLEAR_STATE',
-};
+import types from './mutation-types';
 
 export default {
     [types.SET_ATTRIBUTE_ID](state, id) {
@@ -37,7 +20,7 @@ export default {
     [types.REMOVE_ATTRIBUTE_OPTION_VALUE](state, { languageCode, index }) {
         state.optionValues[languageCode].splice(index, 1);
     },
-    [types.SET_MULTILINGUAL_ATTRIBUTE](state, { isMultilingual }) {
+    [types.SET_MULTILINGUAL_ATTRIBUTE](state, isMultilingual) {
         state.isMultilingual = isMultilingual;
     },
     [types.SET_ATTRIBUTE_OPTION_KEY](state, { index, key }) {
