+++ store/authentication/mutation-types.js
@@ -1,0 +1,5 @@
+export default {
+    SET_USER: 'SET_USER',
+    SET_JWT_TOKEN: 'SET_JWT_TOKEN',
+    CLEAR_STATE: 'CLEAR_STATE',
+};
