+++ store/authentication/mutations.js
@@ -2,11 +2,7 @@
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
-export const types = {
-    SET_USER: 'SET_USER',
-    SET_JWT_TOKEN: 'SET_JWT_TOKEN',
-    CLEAR_STATE: 'CLEAR_STATE',
-};
+import types from './mutation-types';
 
 export default {
     [types.SET_USER](state, user) {
